requirements.txt
asgiref==3.8.1
astroid==3.3.5
colorama==0.4.6
dill==0.3.9
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.3.1
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
